chore: split OrchestrationStrategy symbol to it's own file
