+++ code/backend/batch/utilities/orchestrator/OrchestrationStrategy.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class OrchestrationStrategy(Enum):
+    OPENAI_FUNCTION = "openai_function"
+    LANGCHAIN = "langchain"
+    SEMANTIC_KERNEL = "semantic_kernel"
