+++ code/backend/batch/utilities/orchestrator/Strategies.py
@@ -1,10 +1,4 @@
-from enum import Enum
-
-
-class OrchestrationStrategy(Enum):
-    OPENAI_FUNCTION = "openai_function"
-    LANGCHAIN = "langchain"
-    SEMANTIC_KERNEL = "semantic_kernel"
+from .OrchestrationStrategy import OrchestrationStrategy
 
 
 def get_orchestrator(orchestration_strategy: str):
