+++ code/backend/batch/utilities/orchestrator/__init__.py
@@ -2,7 +2,7 @@ import os
 from typing import List
 import os.path
 import pkgutil
-from .Strategies import OrchestrationStrategy
+from .OrchestrationStrategy import OrchestrationStrategy
 
 
 class OrchestrationSettings:
